implementation of payment api

+++ database/migrations/2017_09_11_183511_alter_request_1_table.php
@@ -14,7 +14,7 @@ class AlterRequest1Table extends Migration
     public function up()
     {
         Schema::table('requests', function (Blueprint $table) {
-            $table->integer('price')->nullable()->after('duration');
+            $table->decimal('price', 5, 2)->nullable()->after('duration');
             $table->boolean('accepted_price')->after('price');
         });
     }
